+++ config/merge_ini.go
@@ -7,8 +7,8 @@ import (
 	"strings"
 	"time"
 
-	"github.com/DataDog/datadog-trace-agent/backoff"
 	"github.com/DataDog/datadog-trace-agent/model"
+	"github.com/DataDog/datadog-trace-agent/writer/backoff"
 	writerconfig "github.com/DataDog/datadog-trace-agent/writer/config"
 	log "github.com/cihub/seelog"
 	"github.com/go-ini/ini"
